Move LoadIterator to separate header and replace include
