+++ thrust/thrust/system/cuda/detail/core/load_iterator.h
@@ -1,0 +1,81 @@
+/******************************************************************************
+ * Copyright (c) 2016, NVIDIA CORPORATION.  All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *     * Redistributions of source code must retain the above copyright
+ *       notice, this list of conditions and the following disclaimer.
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *     * Neither the name of the NVIDIA CORPORATION nor the
+ *       names of its contributors may be used to endorse or promote products
+ *       derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL NVIDIA CORPORATION BE LIABLE FOR ANY
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ ******************************************************************************/
+#pragma once
+
+#include <cub/iterator/cache_modified_input_iterator.cuh>
+
+#include <thrust/type_traits/is_contiguous_iterator.h>
+
+// #include <cuda/std/iterator>
+
+THRUST_NAMESPACE_BEGIN
+
+namespace cuda_cub
+{
+namespace core
+{
+
+// LoadIterator
+// ------------
+// if trivial iterator is passed, wrap loads into LDG
+//
+template <class PtxPlan, class It>
+struct LoadIterator
+{
+  using value_type = typename ::cuda::std::iterator_traits<It>::value_type;
+  using size_type  = typename ::cuda::std::iterator_traits<It>::difference_type;
+
+  using type =
+    ::cuda::std::conditional_t<is_contiguous_iterator<It>::value,
+                               cub::CacheModifiedInputIterator<PtxPlan::LOAD_MODIFIER, value_type, size_type>,
+                               It>;
+}; // struct Iterator
+
+template <class PtxPlan, class It>
+typename LoadIterator<PtxPlan, It>::type _CCCL_DEVICE _CCCL_FORCEINLINE
+make_load_iterator_impl(It it, thrust::detail::true_type /* is_trivial */)
+{
+  return raw_pointer_cast(&*it);
+}
+
+template <class PtxPlan, class It>
+typename LoadIterator<PtxPlan, It>::type _CCCL_DEVICE _CCCL_FORCEINLINE
+make_load_iterator_impl(It it, thrust::detail::false_type /* is_trivial */)
+{
+  return it;
+}
+
+template <class PtxPlan, class It>
+typename LoadIterator<PtxPlan, It>::type _CCCL_DEVICE _CCCL_FORCEINLINE make_load_iterator(PtxPlan const&, It it)
+{
+  return make_load_iterator_impl<PtxPlan>(it, typename is_contiguous_iterator<It>::type());
+}
+
+} // namespace core
+} // namespace cuda_cub
+
+THRUST_NAMESPACE_END
